pprint support: Fix querying middleware
The pprint support erroneously used the value of the middleware as the
map to look into. Also, it seems that the type of the middleware is a
string, not a symbol.

This fixes sthis by swapping the argument order and changing the pretty
printer lookup up to use string keys.