Always prefix error loc-str with a space (#41)
* Always prefix error loc-str with a space

With or without a cause, we need a space between the previous part of
the message and the loc-str. Without this you'll end up with the ex-name
right up against the loc-str. This way we separate the loc-str from the
ex-name if there's no cause while still having a space if there is a
cause.

* Correct nREPL server error string test